backend/financialTips/finacialtips/requirements.txt
requests>=2.25.0
python-dotenv>=1.0.0

